requirements.txt
alembic==1.0.7
aniso8601==4.1.0
attrs==18.2.0
Babel==2.6.0
bcrypt==3.1.6
cffi==1.12.1
Click==7.0
Flask==1.0.2
Flask-Babel==0.12.2
Flask-Bcrypt==0.7.1
Flask-Cors==3.0.7
Flask-Migrate==2.4.0
flask-restplus==0.12.1
Flask-Script==2.0.6
Flask-SQLAlchemy==2.3.2
Flask-Testing==0.7.1
itsdangerous==1.1.0
Jinja2==2.10
jsonschema==3.0.0
Mako==1.0.7
MarkupSafe==1.1.1
pycparser==2.19
PyJWT==1.7.1
PyMySQL==0.9.3
pyrsistent==0.14.11
python-dateutil==2.8.0
python-editor==1.0.4
pytz==2018.9
six==1.12.0
SQLAlchemy==1.2.18
Werkzeug==0.14.1
